requirements.txt
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
Django==4.1
django-coverage-plugin==2.0.3

django-filter==22.1
django-nose==1.4.7
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
flake8==5.0.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.7.0
mwparserfromhell==0.6.4
nose==1.3.7
packaging==21.3
psycopg2-binary==2.9.1
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
wikipedia==1.4.0
